% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ASCIIfy}
\alias{ASCIIfy}
\title{Create an ASCII representation of a string}
\usage{
ASCIIfy(text, font = NULL)
}
\arguments{
\item{text}{A string to be ASCII-fied}

\item{font}{The font to use}
}
\description{
Uses http://artii.herokuapp.com/ API, so requires an active internet connection.
}
\examples{
ASCIIfy( "test" )
}
\seealso{
AvailableFonts
}
