% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{CoolToBeASCII}
\alias{CoolToBeASCII}
\title{Prints a super cool ASCII title in the font of your choice}
\usage{
CoolToBeASCII(text, font = NULL)
}
\arguments{
\item{text}{A string to be ASCII-fied}

\item{font}{The font to use. See http://artii.herokuapp.com/fonts_list for a comprehensive list of available fonts.}
}
\description{
Uses http://artii.herokuapp.com/ API, so requires an active internet connection.
}
\examples{
CoolToBeASCII("Step 1: Do something")
CoolToBeASCII("Step 2: ???", font = "slant" )
CoolToBeASCII("Step 3: Profit", font = "starwars" )
}
\seealso{
AvailableFonts
}
